Fixed angle getters for MotionUpdateEvent

+++ src/main/java/me/zero/client/api/event/defaults/MotionUpdateEvent.java
@@ -139,7 +139,7 @@ public final class MotionUpdateEvent implements Helper {
      *
      * @return The Yaw rotation
      */
-    public double getYaw() {
+    public float getYaw() {
         return nRotation.getX();
     }
 
@@ -148,7 +148,7 @@ public final class MotionUpdateEvent implements Helper {
      *
      * @return The Pitch rotation
      */
-    public double getPitch() {
+    public float getPitch() {
         return nRotation.getY();
     }
